HW15: Refactor handler layer
