+++ hw15_go_sql/internal/server/app/app.go
@@ -16,6 +16,7 @@ import (
 	"github.com/shatilovlex/golang_home_work_basic/hw15_go_sql/internal/config"
 	"github.com/shatilovlex/golang_home_work_basic/hw15_go_sql/internal/repository"
 	"github.com/shatilovlex/golang_home_work_basic/hw15_go_sql/internal/server/handler"
+	"github.com/shatilovlex/golang_home_work_basic/hw15_go_sql/internal/server/handler/shopEndpoint"
 	"github.com/shatilovlex/golang_home_work_basic/hw15_go_sql/internal/usecase"
 	"github.com/shatilovlex/golang_home_work_basic/hw15_go_sql/pkg/pgconnect"
 )
@@ -55,7 +56,7 @@ func (a *App) Start() {
 	flag.Parse()
 
 	ucGetUsers := usecase.NewGetUsersUseCase(querier)
-	endpoint := handler.NewGetUsersEndpoint(ctx, ucGetUsers)
+	endpoint := shopEndpoint.NewGetUsersEndpoint(ctx, ucGetUsers)
 
 	addr := fmt.Sprintf("%v:%v", *ip, *port)
 	server := &http.Server{
@@ -81,10 +82,10 @@ func (a *App) Start() {
 	log.Println("final")
 }
 
-func (a *App) InitMux(endpoint handler.GetUsersEndpoint) http.Handler {
+func (a *App) InitMux(endpoint shopEndpoint.ShopEndpoint) http.Handler {
 	mux := http.NewServeMux()
 
-	mux.HandleFunc("/users", endpoint.GetUsersHandler)
+	handler.MakeShopHandlers(mux, endpoint)
 
 	return mux
 }
